HCKeystore: Create basic hot and cold key stores.
This commit adds an initial implementation of the stateful deterministic
wallets by implementing the underlying algorithms to properly sign and
verify messages (transactions).

For now, wallets are not stored anywhere and they are not integrated
with Ethereum in any way.
The generation of the state and wallet synchronization is not
implemented yet.